Financial Agent/requirements.txt
agno
openai
python-dotenv
duckduckgo-search
lancedb
pypdf
streamlit
pandas
yfinance
tantivy
groq
fastapi
uvicorn